on the contrary, string length is int

+++ ext/bz2/bz2.c
@@ -16,7 +16,7 @@
   +----------------------------------------------------------------------+
 */
  
-/* $Id: bz2.c,v 1.14.2.3.2.12.2.3 2008/06/22 12:03:30 indeyets Exp $ */
+/* $Id: bz2.c,v 1.14.2.3.2.12.2.4 2008/07/01 07:36:59 tony2001 Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include "config.h"
@@ -366,7 +366,7 @@ static PHP_FUNCTION(bzopen)
 {
 	zval    **file;   /* The file to open */
 	char     *mode;   /* The mode to open the stream with */
-	long      mode_len;
+	int      mode_len;
 
 	BZFILE   *bz;     /* The compressed file stream */
 	php_stream *stream = NULL;
@@ -488,7 +488,7 @@ static PHP_FUNCTION(bzcompress)
 					  block_size  = 4, /* Block size for compression algorithm */
 					  work_factor = 0, /* Work factor for compression algorithm */
 					  argc;            /* Argument count */
-	long              source_len;      /* Length of the source data */
+	int               source_len;      /* Length of the source data */
 	unsigned int      dest_len;        /* Length of the destination buffer */ 
 
 	argc = ZEND_NUM_ARGS();
